app/backend/requirements.txt
coverage==5.3
django-cors-headers==3.5.0
django-filter==2.4.0
django-mysql==3.8.1
django-nose==1.4.7
django==3.1.2
djangorestframework==3.12.1
mysql-connector-python==8.0.21
mysqlclient==2.0.1
pycodestyle==2.6.0
wheel==0.35.1
pinocchio==0.4.2
nose==1.3.7
nose-exclude==0.5.0
drf-yasg==1.20.0
